knowledge_server/requirements.txt
fastapi==0.116.1
json_repair==0.49.0
llama_index==0.13.2
pydantic==2.11.7
python-dotenv==1.1.1
qdrant_client==1.15.1
schedule==1.2.2
SQLAlchemy==2.0.43
uvicorn==0.35.0
llama-index-vector-stores-postgres
llama-index-vector-stores-qdrant
pymysql
llama-index-llms-openai-like